+++ network/src/protocols/mod.rs
@@ -35,10 +35,13 @@ pub type PeerIndex = SessionId;
 pub type BoxedFutureTask = Pin<Box<dyn Future<Output = ()> + 'static + Send>>;
 
 use crate::{
-    Behaviour, Error, NetworkState, Peer, ProtocolVersion, SupportProtocols,
-    compress::{compress, decompress},
+    Behaviour, Error, NetworkState, Peer, ProtocolVersion,
+    compress::LengthDelimitedCodecWithCompress,
     network::{async_disconnect_with_message, disconnect_with_message},
 };
+
+#[cfg(not(target_family = "wasm"))]
+use crate::SupportProtocols;
 
 /// Abstract protocol context
 #[async_trait]
@@ -81,6 +84,20 @@ pub trait CKBProtocolContext: Send {
     /// Filter broadcast message
     async fn async_filter_broadcast(&self, target: TargetSession, data: Bytes)
     -> Result<(), Error>;
+    /// Filter broadcast message through queue and specify protocol id
+    async fn async_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error>;
+    /// Filter broadcast message through quick queue and specify protocol id
+    async fn async_quick_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error>;
     /// Disconnect session
     async fn async_disconnect(&self, peer_index: PeerIndex, message: &str) -> Result<(), Error>;
     /// Send message through quick queue
@@ -94,6 +111,13 @@ pub trait CKBProtocolContext: Send {
     fn quick_send_message_to(&self, peer_index: PeerIndex, data: Bytes) -> Result<(), Error>;
     /// Filter broadcast message through quick queue
     fn quick_filter_broadcast(&self, target: TargetSession, data: Bytes) -> Result<(), Error>;
+    /// Filter broadcast message through quick queue and specify protocol id
+    fn quick_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error>;
     /// spawn a future task, if `blocking` is true we use tokio_threadpool::blocking to handle the task.
     fn future_task(&self, task: BoxedFutureTask, blocking: bool) -> Result<(), Error>;
     /// Send message
@@ -174,6 +198,7 @@ pub struct CKBProtocol {
     max_frame_length: usize,
     handler: Box<dyn CKBProtocolHandler>,
     network_state: Arc<NetworkState>,
+    compress: bool,
 }
 
 impl CKBProtocol {
@@ -191,6 +216,7 @@ impl CKBProtocol {
             supported_versions: support_protocol.support_versions(),
             network_state,
             handler,
+            compress: true,
         }
     }
 
@@ -214,7 +240,14 @@ impl CKBProtocol {
                 versions.sort_by(|a, b| b.cmp(a));
                 versions.to_vec()
             },
+            compress: true,
         }
+    }
+
+    /// Enable or disable compression, default is enabled
+    pub fn compress(mut self, enable: bool) -> Self {
+        self.compress = enable;
+        self
     }
 
     /// Protocol id
@@ -245,11 +278,13 @@ impl CKBProtocol {
             .id(self.id)
             .name(move |_| protocol_name.clone())
             .codec(move || {
-                Box::new(
+                Box::new(LengthDelimitedCodecWithCompress::new(
+                    self.compress,
                     length_delimited::Builder::new()
                         .max_frame_length(max_frame_length)
                         .new_codec(),
-                )
+                    self.id,
+                ))
             })
             .support_versions(supported_versions)
             .service_handle(move || {
@@ -259,8 +294,6 @@ impl CKBProtocol {
                     handler: self.handler,
                 }))
             })
-            .before_send(compress)
-            .before_receive(|| Some(Box::new(decompress)))
             .build()
     }
 }
@@ -486,6 +519,28 @@ impl CKBProtocolContext for DefaultCKBProtocolContext {
             .await?;
         Ok(())
     }
+    async fn async_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error> {
+        self.async_p2p_control
+            .filter_broadcast(target, proto_id, data)
+            .await?;
+        Ok(())
+    }
+    async fn async_quick_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error> {
+        self.async_p2p_control
+            .quick_filter_broadcast(target, proto_id, data)
+            .await?;
+        Ok(())
+    }
     async fn async_disconnect(&self, peer_index: PeerIndex, message: &str) -> Result<(), Error> {
         debug!("Disconnect peer: {}, message: {}", peer_index, message);
         async_disconnect_with_message(&self.async_p2p_control, peer_index, message).await?;
@@ -521,6 +576,16 @@ impl CKBProtocolContext for DefaultCKBProtocolContext {
     fn quick_filter_broadcast(&self, target: TargetSession, data: Bytes) -> Result<(), Error> {
         self.p2p_control
             .quick_filter_broadcast(target, self.proto_id, data)?;
+        Ok(())
+    }
+    fn quick_filter_broadcast_with_proto(
+        &self,
+        proto_id: ProtocolId,
+        target: TargetSession,
+        data: Bytes,
+    ) -> Result<(), Error> {
+        self.p2p_control
+            .quick_filter_broadcast(target, proto_id, data)?;
         Ok(())
     }
     fn future_task(&self, task: BoxedFutureTask, blocking: bool) -> Result<(), Error> {