+++ sync/src/utils.rs
@@ -1,15 +1,17 @@
 use crate::{Status, StatusCode};
 use ckb_logger::error;
 use ckb_network::{CKBProtocolContext, PeerIndex, ProtocolId, SupportProtocols};
-use ckb_types::packed::{RelayMessageReader, SyncMessageReader};
+use ckb_types::packed::{self, RelayMessageReader, SyncMessageReader};
 use ckb_types::prelude::*;
-use std::fmt;
-use std::fmt::Formatter;
+use std::{
+    fmt::{self, Formatter},
+    sync::Arc,
+};
 
 /// Send network message into parameterized `protocol_id` protocol connection.
 ///
 /// Equal to `nc.send_message`.
-#[must_use]
+#[allow(dead_code)]
 pub(crate) fn send_message<Message: Entity>(
     protocol_id: ProtocolId,
     nc: &dyn CKBProtocolContext,
@@ -17,6 +19,67 @@ pub(crate) fn send_message<Message: Entity>(
     message: &Message,
 ) -> Status {
     if let Err(err) = nc.send_message(protocol_id, peer_index, message.as_bytes()) {
+        let name = message_name(protocol_id, message);
+        let error_message = format!("nc.send_message {name}, error: {err:?}");
+        error!("{}", error_message);
+        return StatusCode::Network.with_context(error_message);
+    }
+
+    let bytes = message.as_bytes().len() as u64;
+    let item_name = item_name(protocol_id, message);
+    let protocol_name = protocol_name(protocol_id);
+    metric_ckb_message_bytes(
+        MetricDirection::Out,
+        &protocol_name,
+        &item_name,
+        None,
+        bytes,
+    );
+
+    Status::ok()
+}
+
+pub(crate) async fn quick_send_message_async<Message: Entity>(
+    protocol_id: ProtocolId,
+    nc: &Arc<dyn CKBProtocolContext + Sync>,
+    peer_index: PeerIndex,
+    message: &Message,
+) -> Status {
+    if let Err(err) = nc
+        .async_quick_send_message(protocol_id, peer_index, message.as_bytes())
+        .await
+    {
+        let name = message_name(protocol_id, message);
+        let error_message = format!("nc.quick_send_message {name}, error: {err:?}");
+        error!("{}", error_message);
+        return StatusCode::Network.with_context(error_message);
+    }
+
+    let bytes = message.as_bytes().len() as u64;
+    let item_name = item_name(protocol_id, message);
+    let protocol_name = protocol_name(protocol_id);
+    metric_ckb_message_bytes(
+        MetricDirection::Out,
+        &protocol_name,
+        &item_name,
+        None,
+        bytes,
+    );
+
+    Status::ok()
+}
+
+pub(crate) async fn async_send_message<Message: Entity>(
+    protocol_id: ProtocolId,
+    nc: &Arc<dyn CKBProtocolContext + Sync>,
+    peer_index: PeerIndex,
+    message: &Message,
+) -> Status {
+    // ignore Error return, only happens on shutdown case
+    if let Err(err) = nc
+        .async_send_message(protocol_id, peer_index, message.as_bytes())
+        .await
+    {
         let name = message_name(protocol_id, message);
         let error_message = format!("nc.send_message {name}, error: {err:?}");
         ckb_logger::error!("{}", error_message);
@@ -74,7 +137,7 @@ pub(crate) fn metric_ckb_message_bytes(
 /// Send network message into `nc.protocol_id()` protocol connection.
 ///
 /// Equal to `nc.send_message_to`.
-#[must_use]
+#[allow(dead_code)]
 pub(crate) fn send_message_to<Message: Entity>(
     nc: &dyn CKBProtocolContext,
     peer_index: PeerIndex,
@@ -82,6 +145,24 @@ pub(crate) fn send_message_to<Message: Entity>(
 ) -> Status {
     let protocol_id = nc.protocol_id();
     send_message(protocol_id, nc, peer_index, message)
+}
+
+pub(crate) async fn async_send_message_to<Message: Entity>(
+    nc: &Arc<dyn CKBProtocolContext + Sync>,
+    peer_index: PeerIndex,
+    message: &Message,
+) -> Status {
+    let protocol_id = nc.protocol_id();
+    async_send_message(protocol_id, nc, peer_index, message).await
+}
+
+pub(crate) async fn async_quick_send_message_to<Message: Entity>(
+    nc: &Arc<dyn CKBProtocolContext + Sync>,
+    peer_index: PeerIndex,
+    message: &Message,
+) -> Status {
+    let protocol_id = nc.protocol_id();
+    quick_send_message_async(protocol_id, nc, peer_index, message).await
 }
 
 // As for Sync protocol and Relay protocol, returns the internal item name;
@@ -92,9 +173,7 @@ fn message_name<Message: Entity>(protocol_id: ProtocolId, message: &Message) -> 
             .to_enum()
             .item_name()
             .to_owned()
-    } else if protocol_id == SupportProtocols::RelayV3.protocol_id()
-        || protocol_id == SupportProtocols::RelayV3.protocol_id()
-    {
+    } else if protocol_id == SupportProtocols::RelayV3.protocol_id() {
         RelayMessageReader::new_unchecked(message.as_slice())
             .to_enum()
             .item_name()
@@ -153,5 +232,24 @@ fn protocol_name(protocol_id: ProtocolId) -> String {
             error!("send_message got an unknown protocol id: {}", protocol_id);
             "Unknown".to_owned()
         }
+    }
+}
+
+pub(crate) async fn send_block_proposals(
+    nc: &Arc<dyn CKBProtocolContext + Sync>,
+    peer_index: PeerIndex,
+    txs: Vec<packed::Transaction>,
+) {
+    let content = packed::BlockProposal::new_builder()
+        .transactions(txs)
+        .build();
+    let message = packed::RelayMessage::new_builder().set(content).build();
+    let status = async_quick_send_message_to(nc, peer_index, &message).await;
+    if !status.is_ok() {
+        ckb_logger::error!(
+            "send RelayBlockProposal to {}, status: {:?}",
+            peer_index,
+            status
+        );
     }
 }