+++ util/launcher/src/lib.rs
@@ -13,13 +13,14 @@ use ckb_channel::Receiver;
 use ckb_constant::consensus::ENABLED_SCRIPT_HASH_TYPE;
 use ckb_jsonrpc_types::ScriptHashType;
 use ckb_light_client_protocol_server::LightClientProtocol;
-use ckb_logger::info;
-use ckb_logger::internal::warn;
+use ckb_logger::{error, info, warn};
+use ckb_network::Error;
 use ckb_network::{
     CKBProtocol, Flags, NetworkController, NetworkService, NetworkState, SupportProtocols,
     network::TransportType, observe_listen_port_occupancy,
 };
 use ckb_network_alert::alert_relayer::AlertRelayer;
+use ckb_onion::OnionServiceConfig;
 use ckb_resource::Resource;
 use ckb_rpc::{RpcServer, ServiceBuilder};
 use ckb_shared::shared_builder::{SharedBuilder, SharedPackage};
@@ -29,6 +30,7 @@ use ckb_sync::{BlockFilter, NetTimeProtocol, Relayer, SyncShared, Synchronizer};
 use ckb_tx_pool::service::TxVerificationResult;
 use ckb_verification::GenesisVerifier;
 use ckb_verification_traits::Verifier;
+use std::net::{Ipv4Addr, SocketAddr};
 use std::sync::Arc;
 
 const SECP256K1_BLAKE160_SIGHASH_ALL_ARG_LEN: usize = 20;
@@ -270,6 +272,134 @@ impl Launcher {
         }
     }
 
+    /// Start onion service
+    pub fn start_onion_service(&self, network_controller: NetworkController) -> Result<(), Error> {
+        if !self.args.config.network.onion.listen_on_onion {
+            info!(
+                "onion_config.listen_on_onion is false, CKB won't listen on the onion hidden network"
+            );
+            return Ok(());
+        }
+        let onion_config = self.args.config.network.onion.clone();
+
+        if onion_config.onion_server.is_none() && self.args.config.network.proxy.proxy_url.is_none()
+        {
+            info!(
+                "Neither onion_server nor proxy_url is set in the config file, CKB won't listen on the onion hidden network"
+            );
+            return Ok(());
+        }
+
+        let p2p_listen_address: SocketAddr = {
+            match onion_config.p2p_listen_address {
+                Some(p2p_listen_address) => {
+                    let p2p_listen_address: SocketAddr =
+                        p2p_listen_address.parse().map_err(|err| {
+                            Error::Config(format!("Failed to parse p2p_listen_address: {}", err))
+                        })?;
+                    // p2p_listen_address's port shouldn't be 0
+                    if p2p_listen_address.port() == 0 {
+                        return Err(Error::Config(
+                            "[network.onion].p2p_listen_address's port shouldn't be 0".to_string(),
+                        ));
+                    }
+                    p2p_listen_address
+                }
+                None => {
+                    if let Some(port) = {
+                        self.args
+                            .config
+                            .network
+                            .listen_addresses
+                            .iter()
+                            .find_map(|addr| {
+                                use ckb_network::multiaddr::Protocol;
+
+                                let mut iter = addr.iter();
+                                if let (Some(Protocol::Ip4(ip)), Some(Protocol::Tcp(port))) =
+                                    (iter.next(), iter.next())
+                                {
+                                    if ip == Ipv4Addr::new(0, 0, 0, 0)
+                                        || ip == Ipv4Addr::new(127, 0, 0, 1)
+                                    {
+                                        return Some(port);
+                                    }
+                                }
+                                None
+                            })
+                    } {
+                        SocketAddr::new(std::net::IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), port)
+                    } else {
+                        warn!(
+                            "No suitable IPv4 listen address with host 0.0.0.0 or 127.0.0.1 was found in `[network].listen_addresses`."
+                        );
+                        SocketAddr::new(std::net::IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 8115)
+                    }
+                }
+            }
+        };
+
+        {
+            // check tor_controller is listening
+            let tor_controller_addr = onion_config.tor_controller.parse().map_err(|err| {
+                Error::Config(format!("Failed to parse tor_controller address: {}", err))
+            })?;
+            match std::net::TcpStream::connect_timeout(
+                &tor_controller_addr,
+                std::time::Duration::from_secs(2),
+            ) {
+                Ok(_c) => {
+                    info!(
+                        "CKB has confirmed that onion_config.tor_controller is listening on {}, trying to listen on the onion hidden network by the tor_controller",
+                        onion_config.tor_controller
+                    );
+                }
+                Err(_err) => {
+                    error!(
+                        "tor_controller is not listening on {}, CKB won't try to listen on the onion hidden network",
+                        tor_controller_addr
+                    );
+                    return Ok(());
+                }
+            }
+        }
+
+        let onion_service_config: OnionServiceConfig = OnionServiceConfig {
+            onion_private_key_path: onion_config.onion_private_key_path.unwrap_or(
+                self.args
+                    .config
+                    .network
+                    .onion_private_key_path()
+                    .display()
+                    .to_string(),
+            ),
+            tor_controller: onion_config.tor_controller,
+            tor_password: onion_config.tor_password,
+            p2p_listen_address,
+            onion_external_port: onion_config.onion_external_port,
+        };
+        let (onion_service, onion_service_addr) = ckb_onion::onion_service::OnionService::new(
+            self.async_handle.clone(),
+            onion_service_config,
+            network_controller.node_id(),
+        )
+        .map_err(|err| Error::Config(format!("Failed to create onion service: {}", err)))?;
+
+        self.async_handle.spawn(async move {
+            match onion_service
+                .start(network_controller, onion_service_addr)
+                .await
+            {
+                Ok(()) => {}
+                Err(err) => error!(
+                    "CKB failed to start listening on the onion hidden network: {}",
+                    err
+                ),
+            }
+        });
+        Ok(())
+    }
+
     /// Start network service and rpc serve
     pub fn start_network_and_rpc(
         &self,
@@ -314,11 +444,14 @@ impl Launcher {
         if support_protocols.contains(&SupportProtocol::Filter) {
             let filter = BlockFilter::new(Arc::clone(&sync_shared));
 
-            protocols.push(CKBProtocol::new_with_support_protocol(
-                SupportProtocols::Filter,
-                Box::new(filter),
-                Arc::clone(&network_state),
-            ));
+            protocols.push(
+                CKBProtocol::new_with_support_protocol(
+                    SupportProtocols::Filter,
+                    Box::new(filter),
+                    Arc::clone(&network_state),
+                )
+                .compress(false),
+            );
         } else {
             flags.remove(Flags::BLOCK_FILTER);
         }
@@ -420,7 +553,12 @@ impl Launcher {
                     .map(|script| script.clone().into())
                     .collect(),
             )
-            .enable_alert(alert_verifier, alert_notifier, network_controller.clone())
+            .enable_alert(
+                alert_verifier,
+                alert_notifier,
+                network_controller.clone(),
+                shared.clone(),
+            )
             .enable_indexer(
                 shared.clone(),
                 &self.args.config.db,
@@ -431,6 +569,10 @@ impl Launcher {
         let io_handler = builder.build();
 
         let _rpc = RpcServer::new(rpc_config, io_handler, self.rpc_handle.clone());
+
+        if let Err(err) = self.start_onion_service(network_controller.clone()) {
+            error!("Failed to start onion service: {}", err);
+        }
 
         network_controller
     }