+++ util/migrate/src/migrations/add_chain_root_mmr.rs
@@ -29,10 +29,11 @@ impl Migration for AddChainRootMMR {
                     .template(
                         "{prefix:.bold.dim} {spinner:.green} [{elapsed_precise}] [{bar:40.cyan/blue}] {pos}/{len} ({eta}) {msg}",
                     )
+                    .expect("Failed to set progress bar template")
                     .progress_chars("#>-"),
             );
         pbi.set_position(0);
-        pbi.enable_steady_tick(5000);
+        pbi.enable_steady_tick(std::time::Duration::from_millis(5000));
 
         let mut block_number = 0;
         let mut mmr_size = 0;